Use framework for timm models.
