+++ torchbenchmark/models/timm_nfnet/__init__.py
@@ -1,14 +1,8 @@
-import torch
-import timm.models.nfnet
-
-from ...util.model import BenchmarkModel
+from torchbenchmark.util.framework.timm.model_factory import TimmModel
 from torchbenchmark.tasks import COMPUTER_VISION
-from .config import TimmConfig
-from typing import Tuple
 
-class Model(BenchmarkModel):
+class Model(TimmModel):
     task = COMPUTER_VISION.CLASSIFICATION
-    optimized_for_inference = True
 
     # Original train batch size 128, hardware Nvidia rtx 3090
     # Source: https://gist.github.com/rwightman/bb59f9e245162cee0e38bd66bd8cd77f#file-bench_by_train-csv-L147
@@ -18,57 +12,6 @@ class Model(BenchmarkModel):
     DEFAULT_TRAIN_BSIZE = 128
     DEFAULT_EVAL_BSIZE = 128
 
-    def __init__(self, test, device, jit=False,
-                 variant='dm_nfnet_f0',
-                 batch_size=None, extra_args=[]):
-        super().__init__(test=test, device=device, jit=jit, batch_size=batch_size, extra_args=extra_args)
-
-        self.model = timm.create_model(variant, pretrained=False, scriptable=True)
-        self.cfg = TimmConfig(model = self.model, device = device)
-        self.example_inputs = self._gen_input(self.batch_size)
-        self.model.to(
-            device=self.device
-        )
-        if device == 'cuda':
-            torch.cuda.empty_cache()
-
-    def _gen_input(self, batch_size):
-        return torch.randn((batch_size,) + self.cfg.input_size, device=self.device)
-
-    def _gen_target(self, batch_size):
-        return torch.empty(
-            (batch_size,) + self.cfg.target_shape,
-            device=self.device, dtype=torch.long).random_(self.cfg.num_classes)
-
-    def _step_train(self):
-        self.cfg.optimizer.zero_grad()
-        output = self.model(self.example_inputs)
-        if isinstance(output, tuple):
-            output = output[0]
-        target = self._gen_target(output.shape[0])
-        self.cfg.loss(output, target).backward()
-        self.cfg.optimizer.step()
-
-    def set_eval(self):
-        self.model.eval()
-
-    def _step_eval(self):
-        output = self.model(self.example_inputs)
-        return output
-
-    def get_module(self):
-        return self.model, (self.example_inputs,)
-
-    def train(self, niter=1):
-        if self.device == "cuda":
-            raise NotImplementedError("Disable the train test because it causes CUDA OOM on Nvidia T4")
-        self.model.train()
-        for _ in range(niter):
-            self._step_train()
-
-    def eval(self, niter=1) -> Tuple[torch.Tensor]:
-        self.model.eval()
-        with torch.no_grad():
-            for _ in range(niter):
-                out  = self._step_eval()
-        return (out, )
+    def __init__(self, test, device, jit=False, batch_size=None, extra_args=[]):
+        super().__init__(test=test, model_name='dm_nfnet_f0', device=device,
+                         jit=jit, batch_size=batch_size, extra_args=extra_args)
