+++ torchbenchmark/util/framework/timm/timm_config.py
@@ -1,5 +1,3 @@
-# Generated by gen_timm_models.py
-import torch
 import torch.nn as nn
 import dataclasses
 from timm.optim import create_optimizer
